+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -268,6 +268,7 @@ class DataTableRoot<
       sideIndents,
       selectedRows,
       accordionDuration,
+      data: rawData,
     } = this.asProps;
     const { gridTemplateColumns, gridTemplateAreas } = this.gridSettings;
     return {
@@ -303,6 +304,7 @@ class DataTableRoot<
       onSelectRow: this.handleSelectRow,
       getFixedStyle: this.getFixedStyle,
       onCellClick: this.handleCellClick,
+      rawData,
     };
   }