+++ semcore/data-table/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 ### Added
 
 - Ability to set columns into one grouped column.
+- RawData to `CellRenderProps`.
 
 ### Fixed